Workaround pyspark interval issues

+++ listenbrainz_spark/recording_similarity/recording_similarity.py
@@ -30,7 +30,8 @@ def calculate(window_size: int, similarity_threshold: float, time_threshold: int
                      , LEAD(recording_mbid, {idx}) OVER row_next AS mbid1
                      ,    ( artist_credit_id != LEAD(artist_credit_id, {idx}) OVER row_next 
                         AND recording_mbid != LEAD(recording_mbid, {idx}) OVER row_next
-                        AND (LEAD(listened_at, {idx}) OVER row_next) - listened_at <= {time_threshold}
+                        -- spark-sql supports interval types but pyspark doesn't so currently need to convert to bigints
+                        AND BIGINT(LEAD(listened_at, {idx}) OVER row_next) - BIGINT(listened_at) <= {time_threshold}
                        ) AS similar
                   FROM {table}
                  WHERE recording_mbid IS NOT NULL
